+++ spec/mongoid/nested_attributes_spec.rb
@@ -1628,6 +1628,7 @@ describe Mongoid::NestedAttributes do
                           context "when saving the parent" do
 
                             before do
+                              p persisted.atomic_updates
                               persisted.with(safe: true).save
                             end